[query/optimizers] add group by push down
