+++ fusequery/query/src/optimizers/mod.rs
@@ -7,9 +7,13 @@ mod optimizer_limit_push_down_test;
 
 mod optimizer;
 mod optimizer_filter_push_down;
+mod optimizer_group_by_push_down;
 mod optimizer_limit_push_down;
+mod optimizer_util;
 
 pub use optimizer::IOptimizer;
 pub use optimizer::Optimizer;
 pub use optimizer_filter_push_down::FilterPushDownOptimizer;
+pub use optimizer_group_by_push_down::GroupByPushDownOptimizer;
 pub use optimizer_limit_push_down::LimitPushDownOptimizer;
+pub use optimizer_util::OptimizerUtil;
